Improved accuracy of mainLoops so fps is truly uncapped.
Switched from millisecond time to nanoseconds for better accuracy and
unlimited fps.
Removed synchronisation of balls list from Storage and Physics classes
as they are not necessary due to physics only being done in a single
thread.
Fixed multi-thread issue with client using udp when it has not been
setup.